requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-decouple==3.5
pydantic>=1.8.0,<2.0.0

